requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
email-validator==2.1.0
pdf2image==1.17.0
Pillow==10.0.0
requests==2.31.0
base64io==1.0.3
